Predict if an individual is a high earner
This project aims to build models (Logit, KNN, SVM) that predict whether an individual is a high-earner by observing some of their characteristics. The data is extracted from the 1994 Census database and consist of individual characteristics, such as age, education, marital status, occupation, race, sex, and salary, among others. 
The last column of the dataset is a discrete variable, Salary, and contains one of two strings to indicate if the salary was below or above $50,000. Use this variable to make a categorical outcome variable, “HighIncome” that takes values of 0 and 1. 

Use 70% of the data to train your models and the rest as your test sample. Also, use age, Education level (numeric), hours per week, capital gain, and sex as your input features. [Hint: for some of the models you may need to manually encode other categorical variables such as sex as well.] 
a) Build a Logistic Regression predictive model for individual’s income level. Use model predictions on the test sample to create a confusion matrix and calculate the accuracy score, precision, and recall ratios. 
b) Build a K-Nearest Neighbors classification model for individual’s income level. Use model predictions on the test sample to create a confusion matrix and calculate the accuracy score, precision, and recall ratios. 
c) Build a Support Vector Machine classifier for individual’s income level. Use model predictions on the test sample to create a confusion matrix and calculate the accuracy score, precision, and recall ratios. 
d) Rank the models based on their performance. Can you explain the differences in model performances based on each model’s pros and cons?