Loan default prediction using SVM.ipynb
Part A: 
The goal of the first part is to build a predictive model based on Support Vector Machine to predict loan default probability. The data used for this project was gitten from LendingClub. 
 
a. Use the main (unbalanced) sample to train a SVM classifier. Plot the confusion matrix and calculate the prediction evaluation metrics—i.e., accuracy, precision, and recall rates. 
b. Make a balanced sample by under-sampling the majority class and train a second SVM classifier on this new balanced sample. Show the model performance by calculating the confusion matrix and evaluation metrics. 
c. How did the change in the sample balance affect the prediction performance? 

Part B: Use decision trees, Random Forest, Gradient boosting to predict high earner given the rest of the details
One of the common problems with individual predictive models is that they are prone to overfitting. We want to build ensemble models to decrease the change of overfitting and predict an individual’s income class. These data are extracted from the 1994 Census database and consist of individual characteristics, such as age, education, marital status, occupation, among others. The last column of the dataset is a discrete variable, Salary, and contains one of two strings to indicate if the salary was below or above $50,000. Use this variable to make a categorical outcome variable, “HighIncome” that takes values of 0 and 1. 

1.	Build a Decision Tree classification model for individual’s income level. Use model predictions on the test sample to create a confusion matrix and calculate the accuracy score, precision, and recall ratios. 
2.	Build a Random Forest predictive model for individual’s income level. Use model predictions on the test sample to create a confusion matrix and calculate the accuracy score, precision, and recall ratios. 
c. Build a Gradient Boosting predictive model for individual’s income level. Use model predictions on the test sample to create a confusion matrix and calculate the accuracy score, precision, and recall ratios. 
d. Compare the performance of the Decision Tree model with each of the ensemble models. Can you explain the differences based on each model’s objectives?